+++ components/Prefectures/Prefectures.tsx
@@ -5,7 +5,9 @@ import { PrefecturesItem } from "./PrefecturesItem";
 export const Prefectures: FC<{
   prefCode: number;
   apikey: string;
-}> = ({ prefCode, apikey }) => {
+  handleCheckboxChange: any;
+  selectedPrefs: any;
+}> = ({ prefCode, apikey, handleCheckboxChange, selectedPrefs }) => {
   const [prefs, setprefs] = useState<{ prefCode: string; prefName: string }[]>(
     []
   );
@@ -25,24 +27,6 @@ export const Prefectures: FC<{
       setLoading(false);
     })();
   }, [prefCode, apikey]);
-
-  // 選択された都道府県コードを管理するための状態
-  const [selectedPrefs, setSelectedPrefs] = useState<string[]>([]);
-
-  // チェックボックスの変更を処理する関数
-  const handleCheckboxChange = (code: string) => {
-    console.log("code: ", code);
-    console.log("hoge");
-    setSelectedPrefs((currentSelected) => {
-      if (currentSelected.includes(code)) {
-        // 既に選択されている場合は除去
-        return currentSelected.filter((selectedCode) => selectedCode !== code);
-      } else {
-        // 選択されていない場合は追加
-        return [...currentSelected, code];
-      }
-    });
-  };
 
   return (
     <div>
